Cleaning code and end of game message
Remove un-necessary comments from the code, and re-organize things
so that all the functions are on the bottom, and the main executable
is at the top.

Add a line of code to tell the player thanks for playing, and display
the final scores one last time before ending the program.